Split out LociSet Serializer tests
